@@ -19,8 +19,8 @@
     "@melt-ui/pp": "0.3.2",
     "@melt-ui/svelte": "0.86.0",
     "@sveltejs/adapter-static": "3.0.6",
-    "@sveltejs/kit": "2.7.4",
-    "@sveltejs/vite-plugin-svelte": "3.1.2",
+    "@sveltejs/kit": "2.8.0",
+    "@sveltejs/vite-plugin-svelte": "4.0.0",
     "@tanstack/svelte-query": "5.59.17",
     "autoprefixer": "10.4.20",
     "eslint": "9.14.0",
@@ -31,7 +31,7 @@
     "prettier": "3.3.3",
     "prettier-plugin-svelte": "3.2.7",
     "rollup-plugin-license": "3.5.3",
-    "svelte": "4.2.19",
+    "svelte": "5.1.13",
     "svelte-check": "4.0.5",
     "sveltekit-superforms": "2.20.0",
     "tailwindcss": "3.4.14",
