@@ -1,23 +1,40 @@
 <script lang="ts">
   import IconLoading from '~icons/mdi/loading';
+  import type { Snippet } from 'svelte';
+  import type { HTMLButtonAttributes } from 'svelte/elements';
 
-  export let isLoading = false;
-  export let isDisabled = false;
+  type Props = {
+    isLoading?: boolean;
+    isDisabled?: boolean;
+    class?: string;
+    leftIcon?: Snippet;
+    children?: Snippet;
+    rightIcon?: Snippet;
+  } & HTMLButtonAttributes;
 
-  let className = '';
-  export { className as class };
+  let {
+    isLoading = false,
+    isDisabled = false,
+    class: className = '',
+    leftIcon,
+    children,
+    rightIcon,
+    ...rest
+  }: Props = $props();
+
+  const children_render = $derived(children);
 </script>
 
-<button class={`${className ?? ''}`} {...$$restProps} disabled={isDisabled || isLoading}>
+<button class={`${className ?? ''}`} {...rest} disabled={isDisabled || isLoading}>
   <span class="flex gap-2 justify-center items-center">
-    {#if $$slots.leftIcon}
-      <slot name="leftIcon" />
+    {#if leftIcon}
+      {@render leftIcon?.()}
     {/if}
 
-    <slot />
+    {@render children_render?.()}
 
-    {#if $$slots.rightIcon && !isLoading}
-      <slot name="rightIcon" />
+    {#if rightIcon && !isLoading}
+      {@render rightIcon?.()}
     {/if}
 
     {#if isLoading}
