@@ -6,20 +6,22 @@
   import TagPicker from '$lib/components/TagPicker.svelte';
   import type { Tag } from '$lib/types/tag';
   import { createQuery } from '@tanstack/svelte-query';
-  import { createEventDispatcher } from 'svelte';
+
+  type Props = {
+    defaultTagSet: Array<string>;
+    onapplied: (event: { tag_ids: string[] }) => void;
+  };
+  let { defaultTagSet, onapplied }: Props = $props();
 
   const tagsQuery = createQuery<Array<Tag>>({
     queryKey: [queryKeys.tag.list],
     queryFn: () => getTags({ fetch }),
   });
 
-  $: tags = $tagsQuery.data;
-  export let defaultTagSet: Array<string>;
-
-  const dispatch = createEventDispatcher();
+  const tags = $derived($tagsQuery.data);
 
   function onDelete(id: string) {
-    dispatch('applied', {
+    onapplied({
       tag_ids: defaultTagSet.filter((tagId) => tagId !== id),
     });
   }
@@ -36,7 +38,7 @@
         <SingleTag
           name={tags.find((t) => t.id === id)?.name ?? ''}
           canDelete={true}
-          on:click={() => {
+          onclick={() => {
             onDelete(id);
           }}
         />
@@ -45,8 +47,8 @@
     <div class="mt-6 w-full text-right">
       <TagPicker
         tags={tags.filter((tag) => !defaultTagSet.includes(tag.id))}
-        on:select={({ detail: { tagID } }) => {
-          dispatch('applied', { tag_ids: [...defaultTagSet, tagID] });
+        onselect={({ tagID }) => {
+          onapplied({ tag_ids: [...defaultTagSet, tagID] });
         }}
       />
     </div>
