@@ -6,13 +6,15 @@
   import PageLoadingState from '$lib/components/page-states/PageLoadingState.svelte';
   import PageErrorState from '$lib/components/page-states/PageErrorState.svelte';
 
-  $: query = createQuery<string>({
-    queryKey: [queryKeys.user.self],
-    queryFn: () => getSelf({ fetch }),
-    refetchOnMount: true,
-    staleTime: 0,
-    gcTime: 0,
-  });
+  let query = $derived(
+    createQuery<string>({
+      queryKey: [queryKeys.user.self],
+      queryFn: () => getSelf({ fetch }),
+      refetchOnMount: true,
+      staleTime: 0,
+      gcTime: 0,
+    }),
+  );
 
   const backURL = `/recipes?${localStorage.getItem('last-recipes-query') ?? ''}`;
 </script>
